pyproject.toml
[project]
name = "bilimusic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bilibili-api-python>=17.4.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pywebview>=5.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
